+++ NEWS.md
@@ -1,3 +1,8 @@
+# altdoc 0.2.1
+
+* Fix test failures on CRAN due to the new version of `usethis` 
+  (see https://github.com/cynkra/fledge/issues/683).
+
 # altdoc 0.2.0 
 
 ### Breaking changes